+++ CHANGELOG.md
@@ -50,6 +50,9 @@ changes.
 - Prevent transactions from being resubmitted for application over and over [#485](https://github.com/input-output-hk/hydra-poc/issues/485)
 
 - Update [hydra-node-options](https://hydra.family/head-protocol/docs/getting-started/quickstart/#hydra-node-options) section in docs.
+- Verify cmd line options for hydra node and error out if:
+   - loaded number of peers is exceeded (currently 4)
+   - number of loaded cardano and hydra keys is not the same
 
 ## [0.7.0] - 2022-08-23